feat(formcontrol): add validation and blur handle
