+++ src/V3/StateManagement/Observable.ts
@@ -48,6 +48,7 @@ export class ObservableForm<T extends {}> implements Observable<T> {
    }
 
    notify() {
+      console.log('notify', this.state)
       this.observers.forEach(observer => observer(this.state))
    }
 }
